feat: shared 페이지 검색 기능 구현
