+++ src/pages/Folder/Folder.tsx
@@ -10,6 +10,7 @@ import AddLinkContainer from './components/AddLinkContainer';
 import FoldersContainer from './components/FoldersContainer';
 import NoLinkView from './components/NoLinkView';
 import { CardProps } from '@components/CardsContainer/CardsContainer';
+import filterLinks from '@utils/filterLinks';
 
 function Folder() {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -107,17 +108,3 @@ function Folder() {
 }
 
 export default Folder;
-
-function filterLinks(links?: CardProps[], keyword?: string) {
-  if (!links) return;
-  if (!keyword) return links;
-
-  const lowered = keyword.toLowerCase();
-  return links?.filter(({ url, title, description }) => {
-    return (
-      url?.toLowerCase().includes(lowered) ||
-      title?.toLowerCase().includes(lowered) ||
-      description?.toLowerCase().includes(lowered)
-    );
-  });
-}
