+++ src/pages/Shared/components/Folder/Folder.tsx
@@ -1,10 +1,12 @@
 import * as S from './Folder.style';
+import { useSearchParams } from 'react-router-dom';
 import useRequest from '@hooks/useRequest';
 import { DEFAULT_USER_ID } from '@apis/config/default';
 import SearchBar from '@components/SearchBar';
 import CardList from '@components/CardsContainer';
 import Banner from '../Banner';
 import { CardProps } from '@components/CardsContainer/CardsContainer';
+import filterLinks from '@utils/filterLinks';
 
 interface SharedFolder {
   count: number;
@@ -31,12 +33,16 @@ function Folder() {
   const owner = folder?.owner;
   const links = folder?.links;
 
+  const [searchParams] = useSearchParams();
+  const initialKeyword = searchParams.get('keyword') ?? '';
+  const filteredLinks = filterLinks(links, initialKeyword);
+
   return (
     <main>
       {data && <Banner name={name} owner={owner} />}
       <S.ContentContainer>
         <SearchBar />
-        {data && <CardList cards={links} userId={DEFAULT_USER_ID} />}
+        {data && <CardList cards={filteredLinks} userId={DEFAULT_USER_ID} />}
       </S.ContentContainer>
     </main>
   );
